:recycle:refactor: postMessagePage.js 코드 수정 및 기능을 나눠서 apis 폴더에 파일 생성
